Rename file with a bad name
