+++ envergo/petitions/regulations/ep.py
@@ -10,7 +10,7 @@ from envergo.moulinette.regulations.ep import (
     EspecesProtegeesRegimeUnique,
     EspecesProtegeesSimple,
 )
-from envergo.moulinette.regulations.regime_unique import (
+from envergo.moulinette.regulations.utils import (
     build_ru_hedge_detail_rows,
     collect_zone_configs,
 )
